fix: dev: Fix the buffer overflow in sw_get_info()@ovsdb_if.c.
sysd just buffer overflows when the git branch name is longer than
usual.  This is due to the fact that it blindly strcpy() the branch name
into the switch_version buffer locally.  The fix limits the length of
the switch version name to the certain length, 80 length, and drops the
rest.

Change-Id: Ic4a733d24774daf26616a1fd4d3d060b66115ae2
Signed-off-by: Kei Nohguchi <kei.nohguchi@hpe.com>