chg: dev: changes to set the default VRF status to ready from sysd.
This commit includes changes to set the default VRF status to ready
from sysd. Previous to this commit the default VRF is created by the
sysd but status is set to ready by vrfmgrd. This will cause issues in
reboot scenarios where the sysd will be up and creates the default VRF
but cannot be used till the vrfmgrd sets it to ready.
The creation of namespace and setting the namespace to ready once the
creation of namespace is done should be done at the same place. This
commit ensures that.

Change-Id: I0de35ae8d478880249e9221519c95a26fc095293
Signed-off-by: Jayachandra Babu K <jayachandra.bab.k@hpe.com>